Refactor dmsr_gan module
